device: return singleton instance on require()
Let Lua store the singleton instance in the package.loaded table.

If GPU support is not available or not enabled, preset package.loaded to
an empty table. Detect availability of GPU using a property device.gpu
instead of device().

This makes the HALMD script agnostic with regard to the hardware.

Remove options --disable-gpu and --devices from liquid script.

Add option --disable-gpu to main program.

The CUDA device may be selected with nvlock as an example.