Pass module constructor to module() function
There is no need for passing a module name to module(), since a
module’s canonical name is defined by its path in the file system.

Revise module() to accept a constructor function instead. This avoids
an arbitrary function name such as __new, and ensures that a module
always has a constructor.

The wrapper function around a module constructor uses a tail call.

http://www.lua.org/manual/5.2/manual.html#3.4.9

This reduces the call stack by one level, and, in case of an error,
allows a constructor to use error(..., 2) instead of error(..., 3)
to indicate the call site, which is consistent with error(..., 2)
in non-constructor methods.