thermodynamics: revert from particle_group to particle
There is no point in using a particle_group in thermodynamics.

For direct access to particle, which is ordered by a space-filling curve,
the implementation of selection of a subset of particles would involve
non-coalesced access to the particle arrays, or access via a texture.

Instead of repeating this particle selection in every observable module,
we will implement a method for granular copying from a space-ordered
array in particle to a tag-ordered (sub-)array, or vice versa.