phase_space: prefer simple and precise interface of Lua module
Remove internal loop over particle groups in phase_space module,
which is a seemingly nice, but really horrible idea. Further
require a particle group, and not a particle instance.

This clearly defines the input and output data of the module,
and avoids mistakes such as the loop in the liquid script
over an assumed sequence, which is in fact an object.