blocking_scheme: fix non-zero step origin
The previous implementation assumed that clock->step() is zero at
construction of blocking_scheme. This is not the case for a series
of multiple runs, e.g. with measurement after equilibration.

Initialise block level origins to clock->step() in constructor, and
shift current step by block level origin in the interval comparison.