trajectory: adapt H5MD trajectory file test to fixed array
The h5xx library, and the H5MD readers and writers on top, have a
fatal design flaw: they require containers instead of iterators.
As a consequence, it is not possible determine the size of a dataset
before reading it, and further the h5xx read_dataset methods require
a resizable container.

To work around this design issue, pass a std::vector to the reader,
and after reading copy to the position/velocity sample array. This
will allow use of a fixed-size array type for the sample arrays.