particle: revise iterator interface
The interface of the get_* and set_* methods was inconsistent, since the
former expected only a first iterator, while the latter expected a first
and a last iterator. This was due to an implementation detail, std::copy
requiring the last iterator of the source.

Revise iterator interface to consistently require a first iterator only.
Further, return the last iterator after copying, similar to algorithms
such as std::copy.

In gpu::particle, call cuda::memset for set_position and set_velocity
only, and only if using double-single precision. The other arrays are
initialised to zero in the constructor (e.g. tags, reverse tags), so
there is no need to reset ghost particles in set_* methods.