utility/cache: implement cached value with unit test
The class provides a caching wrapper around an arbitrary value.
The state of the cache is maintained with an internal counter.
A write access to the value increments the counter, while a
read access does not change the counter.

The observer supports transitive dependencies between cached values,
e.g. a chain of caches A → B → …. The equality operator of the cache
observer (of A) enforces the retrieval of the observed cache (A), by
calling a method that computes (from B) the value (A). This method may
in turn compare its own cache observer (of B) to the observed value (B),
which in turn enforces the retrieval of the observed cache (B), etc.