trunc/local_r4: log smoothing parameter
Add module documentation to local_r4, including formulas and plots.

In Lua constructor of lennard_jones potential, store logger instance,
which has a prefix describing the potential and particle instances.

In Lua constructor of local_r4 module, store smoothing parameter as
instance attribute, and add method log() which outputs the parameter
to a given logger.

In Lua constructor of pair_trunc, if a smoothing function is specified,
call its log() method and pass the logger instance of the potential.

Note that the C++ object itself cannot return the smoothing parameter,
since for efficiency reasons (the functor is copied to the GPU as an
argument) it only stores a derived floating-point value (1 / h^2).