permute: add unit test for array permutation on host
This algorithm is used on the host to permute particle arrays given a
permutation array, e.g. particle indices ordered by a space-filling
curve.

Ensure that halmd::permute is passed random access iterators, and that
the index iterator is of integer value type. Add assert() to debug
out-of-bound permutation indices.