boltzmann: fix missing factor of 2 for E_kin in v_scaled
Thanks to Nicolas Höft for pointing this out.

Further, in the formula for v_scaled, move the dimension d behind the
temperature T*, to avoid typesetting confusion with differential dT.

Use brackets around v_cm in square of the velocity of the centre of
mass, to avoid confusion with squared velocity of the centre of mass.