thermodynamics: track input arrays for accurate caching
The previous implementation tracked the step of the clock, and cached
values were updated only if the step changed. This fails to work with
C++ unit tests and Lua scripts, since the input arrays may change while
the step remains the same.

Use halmd::cache observers to track input arrays for changes. Use the
convenient comparison operator of std::tuple to track multiple input
arrays with a single statement.