force: create interface force with unit test
Add base classes force for host and GPU, which provide an interface to
arrays for forces and auxiliary variables (e.g. potential energy).
This reverts to and extends the design of the 0.2.x branch, which is
better suited for caching and parallel execution.

Move get_* methods from particle to force, and write separate unit test.

Note the force array has been renamed to net_force, since the name
conflicts with the name of the class, i.e. the constructor name.