gpu/reduce: support passing constant value using generic iterator
Rewrite reduce kernel for use with generic iterators. This allows
reducing multiple input arrays (e.g. in correlations functions)
using halmd::zip_iterator, or passing an arbitrary constant value
using halmd::constant_iterator with halmd::zip_iterator.

The number of used registers for a reduce kernel remains the same, since
only the first iterator is passed to the kernel function, together with
the number of elements. This avoids a needless copy of the second and
subsequent last iterators contained in a zip iterator.

Note that the number of elements is passed as std::ptrdiff_t, which
corresponds to 32-bit on 32-bit machines, and 64-bit on 64-bit machines.
The number of registers for the binary output for sm_20 (Fermi) remains
the same, while the number of registers in the binary code for sm_13
(pre-Fermi) is slightly increased for 64-bit GPUs in some cases.

constant_iterator implements a partial Random Access Iterator that returns
a constant value when accessed at any given offset. In particular, the
iterator does not hold a distance, which avoids a duplicate distance
value when combined with a Random Access Iterator using a zip iterator.

zip_iterator combines multiple Random Access Iterators into a partial
Random Access Iterator. The iterator bridges the gap between the host,
where a std::tuple of iterators is used for construction, and the GPU,
where a halmd::tuple of value references is returned upon dereference
at an offset.

Extend reduce unit test to cover possible combinations of iterators.

Adapt thermodynamics, verlet_nvt_hoover and tagged_particle.