thermodynamics test: reduce sampling rate
The imposed test tolerances are based on the measured error of mean, which is
only reliable if the individual samples are uncorrelated. Thus we increase the
sampling rate to approximately match the relaxation time of the fluid (obtained
from the initial decay of the VACF for d=3).

This commit addresses the constantly failing unit tests and fixes #81, #82.