packages.mk: consistently link git-annex against static pcre library
Replace CPATH and LIBRARY_PATH with --extra-include-dirs and --extra-lib-dirs,
since the former could be overriden by a system-wide shared pcre library.

http://www.haskell.org/cabal/users-guide/installing-packages.html#miscellaneous-options

This ensures that git-annex depends only on core system libraries, which
allows the user to copy the git-annex binary to other machines that do not
have (up-to-date versions of) the pcre or gmp library.