implement Lennard-Jones potential with a linear cutoff
The linear cutoff modifies the potential drastically and, actually, we do not
recommend it for future work. It is mainly provided for historical reasons to
connect to existing data and publications.

The linear cutoff can not be implemented via the current interface of the
forces::trunc functors since the cutoff parameters require knowledge about the
potential parameters. Hence, a separate potential module has been created.