gpu/phase_space: fix initialisation from multiple particle groups
When multiple particle groups are used (e.g. one per species) for
initialisation, phase_space::set() needs to preserve the existing
particle data for particles not part of the group.

Note that phase_space::set() is inefficient, since it copies all
particles to and from the host. It will soon be superseded by
efficient free functions to initialise particles given a particle
group.