examples/liquid: add scripts for Kob-Andersen mixture
Add scripts to equilibrate and observe a supercooled Lennard-Jones
mixture with potential parameters σ and ε by Kob and Andersen.

http://dx.doi.org/10.1103/PhysRevLett.73.1376
http://dx.doi.org/10.1103/PhysRevE.51.4626
http://dx.doi.org/10.1103/PhysRevE.52.4134

This script gives an example of using multiple particle groups with one
group per species, with trajectories written to and read from separate
groups in the H5MD file.

The default values for density and temperature are equivalent to those
for the benchmark of a Kob-Andersen mixture in the HALMD documentation.