integration/continuation: compare H5MD trajectory in continuation tests
Add test case that compares two H5MD trajectory groups given a pair of
H5MD files and dataset offsets. The comparison is binary exact, which
assumes reproducible floating-point operations used in the periodic
extension of particle positions for writing.

This test uncovers an error in resuming from an H5MD trajectory in the
case of a mixture, which has been discovered thanks to Felix Höfling.
This affects only the 0.2.x version, and is fixed in the next commits.

The following tests passed:
        integration/continuation/lennard_jones/host/2d
        integration/continuation/lennard_jones/host/3d
        integration/continuation/lennard_jones_simple/host/2d
        integration/continuation/lennard_jones_simple/host/3d
        integration/continuation/lennard_jones_simple/gpu/2d
        integration/continuation/lennard_jones_simple/gpu/3d

The following tests FAILED:
        integration/continuation/lennard_jones/gpu/2d (Failed)
        integration/continuation/lennard_jones/gpu/3d (Failed)