Fix duplicate tags when resuming from multiple trajectory groups.
Commit:e69c24c assumed that particle tags are local to a species,
which is not the case. A particle tag uniquely identifies a particle
within an instance of host::particle or gpu::particle.

This broke reverse tags, since the index of the g_reverse_tag array
is defined as the particle tag. This caused disordered trajectories
upon resuming from multiple trajectory groups, e.g. a mixture.

This partially fixes #75.