packages.mk: configure halmd with CUDACC="nvcc --compiler-bindir=/usr/bin"
This is needed if the CUDA compiler does not support the GCC compiler
configured in PATH, e.g. when using GCC 4.7 and NVCC ≤ 4.2 on a system
with an older GCC version in /usr/bin.

Note that a better, long term practical solution is to ask your system
administrator to add the line “compiler-bindir = /usr/bin” to the file
<cuda>/bin/nvcc.profile.