thermodynamics test (2d): add self-made reference values
So far, basically no thermodynamic quantities were tested for d=2 and it is
better to test against self-made references at least. Of course, it would be
nicer to have reference values from an external source.

Further, the tolerances on the specific heat and the compressibility are
adjusted. The latter in particular is very difficult to measure to a reasonable
precision.

This picks commit:4062cc9 from branch 0.2.x and fixes #93.