enable shifting and rescaling of initial particle velocities
Shifting of the initial velocities is needed to set the center-of-mass velocity
to zero if one continues from an equilibration run that was done with the
Andersen thermostat.

Rescaling of the velocities is very useful (even crucial?) for proper
equilibration in the NVE ensemble at a given temperature.

The functionality was already implemented, but not accessible via the Lua
interface.