thermodynamics test: refine tolerances on results
The test failed occassionally for the response coefficients, i.e., the specific
heat and the compressibility. The tolerances of the state variables (pressure,
potential energy) were choosen as 4.5σ based on the width of their
distribution. For the response coefficients, these standard deviations are not
easily obtained during the test run, and so far, ad hoc values were used. This
is fixed here by determining the distributions of c_V and χ_S from 100 test
runs. As a result, the respective tolerances increase somewhat.

Some other checks failed as well, albeit rarely. Tolerances were improved using
the width of the distribution for each observable (measured from at least 100
test runs) instead of the error of mean obtained during a single test run.

The tolerances on momentum and energy conservation are still ad hoc and were
slightly adjusted.

The revised test has successfully passed 1000 consecutive runs.

This commit fixes #93.