packages.mk: fix LuaJIT options -jv and -jdump
These features require the LuaJIT jit modules, which previously could
not be loaded due to an invalid package.path. Remove the VERSION flag
from compilation, as the release version is hard-coded in the source.
Set PREFIX not only during installation, but also during compilation,
to generate the correct package.path.