liquid scripts: add options for cutoff and smoothing
The liquid scripts are still very basic and need some improvements towards a
better user experience. This is just a very first step.

In addition to the new options, allow observables to be deselected by passing
zero sampling interval.

Adjust default values for particle number to a GPU-relevant scale (10,000) and
reduce equilibration timestep to 0.005.