packages.mk: fix installation prefix of Git Perl module
When invoking make as a sub-command of make, command-line variables
such as PREFIX are passed to the sub-command in the environment
variables MAKEFLAGS and PREFIX. This breaks installation of the
Git Perl module, since its makefile ignores --prefix and installs
to PREFIX, the parent directory of --prefix.

This commit fixes git send-email, which uses the Git Perl module.