Add Velocity Verlet integrator with Boltzmann distribution
This integrator combines the velocity-Verlet algorithm and the Boltzmann
velocity distribution. At a periodic interval, the velocities are assigned
from a Boltzmann velocity distribution in the second integration half-step.

This integrator is especially useful for cooling or heating a system to a
nominal temperature. After an assignment from the Boltzmann distribution,
the centre of mass velocity is shifted to exactly zero, and the velocities
are rescaled to exactly the nominal temperature.