Instantiate pair_full and pair_trunc templates in potential modules
There are two choices for the compilation of forces and potentials:
Either forces (pair_trunc, pair_full) are modular and potentials
non-modular, or forces are non-modular and potentials modular.
Given the larger number of potentials, the latter is preferable.

The file size remains the same, e.g., 18316 kB before this commit,
18228 kB after this commit, thus there is no advantage to compiling
all instantiations in the same translation unit.