Convert binning and neighbour lists to data-driven interfaces
With the old C++ program, the course of the simulation was rigid.
The program iterated for N steps, and every observable would be
calculated at most N times.

The new script interface makes it possible to evaluate observables
between sampler runs as part of the simulation, e.g. to rescale to
a nominal temperature according to the current temperature. This
is implemented elegantly using a data-driven approach.

Make the update() methods of binning and neighbour implementations
private functions, and automatically update the cell or neighbour
lists if necessary. The cell lists are updated if the positions
have changed.

The neighbour lists are updated if the particle identities changed
(e.g., after particle sorting), or if the maximum displacement of
all particles exceeds half the skin distance.

Particle sorting is connected to the on_prepend_update signal
of neighbour. Otherwise, no signal connections are needed.

In neighbour and binning, define array_type.

Adapt header order and remove using namespace.

Shorten maximum_squared_displacement to max_displacement.

Note that std::pair for some constructor parameters of neighbour
implementations is necessary to work around the Luabind limit of
10 arguments.

In the thermodynamics and verlet_nvt_hoover tests, remove core,
which is no longer needed thanks to the data-driven interfaces.