update benchmarks using release version 0.2.0
Modify Lennard-Jones benchmark parameters: increase density to 0.4 (as it was
before), cutoff to 3σ and temperature to 1.2 for comparison with other
simulation software and the neighbour list size to 0.7 (which turned out to be
slightly more efficient).

Improve benchmarking scripts. For the transition to the scripting interface in
master, it should suffice to replace the .rc files by Lua scripts.

(Almost) disable evaluation of observables.