pair_trunc: read first particle by coalescable fetch
The position of the first ("i") particle of the force loop is fetched in linear
order allowing for a coalescable memory transaction. Read directly rather than
via the texture cache.

The change improves the force kernel by about 10% in the Lennard-Jones
benchmark. (Execution time decreases from 2.4 to 2.2 ms on a card of the S1070
node.)

This regression was introduced in commit b8c8001.