forces: compute full stress tensor
Before, only the trace and the off-diagonal elements of the stress tensor were
stored. For non-isotropic systems, however, the diagonal elements themselves
are off interest as well.

On the GPU, the stress tensor being symmetric has 6 entries in 3 dimensions.
For coalescable memory access, they are stored in two arrays of float4 and
float2. In 2 dimensions, float2 and float are used.

On the host, the stress tensor contribution is stored particle-wise as on the
GPU. Thereby, the stress tensor can be evaluated locally. Note that there is a
conceptual problem if the density is (locally) inhomogeneous; see
Varnik, Baschnagel & Binder, J. Chem. Phys. 113, 4444 (2000) [http://dx.doi.org/10.1063/1.1288390]
for a enlightening discussion of the issue.