observables/phase_space: specify which fields shall be read or written
This allows the user to select which data are stored in and fetched from
the /trajectory group of the H5MD output file.

Thereby, storing masses and species in the liquid examples can be
dropped in the future. A missing prerequisite is the revision of
phase_space:set() along with the splitting of the phase_space sample
into individual arrays. Currently, the mass and species values are
overwritten by phase_space:set(sample) and there is no method to restore
them to a constant value.