Calculate the full symmetric stress tensor
Previously only the trace and diagonal elements were stored limiting the
use of the stress tensor to isotropic situations.

In order to allow coalesced access of the stress tensor in the gpu
implementation, the matrix elements are stored in column-major order.
This, on the other hand, creates the need for special reading and
writing functions for conversation between column- and row-major orders.