fix timestep in continuation tests
The timestep was erroneously increased from 0.002 to 0.02 in
commit:f8b5204. The large value of 0.02 lead to divergent trajectories
and non-sense particle positions (~1e34) in the H5MD file of the
equilibration run. Then, the continuation was trapped in an (almost)
infinite loop mapping back such huge values to the box.

Further, the default timestep in the example scripts is reduced to 0.005
(which still appears to be quite large).

In branch 0.2.x, an exception was thrown upon a diverging potential
energy. This check was silently removed in commit:3da56f0.

I thank Nicolas Höft for debugging the issue and pointing out a
solution.

This refs #166789.