integrators: fix order of cache accesses
In commit:8b5ecb6f the order of cache accesses was changed
inadvertently. This order is crucial since commit:bb52a29.

Accessing the cached particle data (velocity, position) with write
permissions invalidates the caches. Accessing the force afterwards has
the side effect that it is recalculated since it depends on the
"current" position. Even worse, the cache observer of the position
stored within the force class already identifies the "new" cache (but
with the old data). Thus, the force is not updated on the following
read—completely breaking the integration scheme!

Add an alert comment everywhere.

It seems that a useful rule is to obtain write access only after read
access. Apply this rule to the phase_space and binning modules as well.

I thank Nicolas Höft for detecting the issue.

This apparently fixes #166921.