observables/thermodynamics: revise interface and documentation
Change names of Lua bindings to reflect the group names of the H5MD
output files. This follows the coding policy to avoid abbreviations in
interface definitions and it provides a 1:1 mapping between field names
and module methods.

Change interface of writer() analogously to phase_space:writer() such
that the user may select which observables are evaluated and written.
Provide defaults for the most commonly used fields that comply with the
H5MD specification. (Note that the total energy may be obtained from
potential_energy and temperature.)

Extend the documentation and provide formulas for each observable.