lua: add defaults to M.add_options()
The user interface shall have full control over default values for the
command line options:
  i) set values
  ii) either disable the default or fall back to a hard-coded value

The defaults are provided as an optional dictionary, omitting the
argument (or a certain key) enables either hard coded defaults or makes
the respective option mandatory. This works nicely because Lua returns a
missing key as `nil` and assigning `nil` to a dictionary key deletes the
entry.