observables.ssf: require normalisation factor from the user
Drop automatic calculation of the normalisation factor. In a
multi-component mixture, it is typically the total number of particles.
As such it cannot be inferred from the particle count of just 2 density
modes. Further, one has to distinguish wether these density modes refer
to disjoint particle subsets or to the same particle group.

Fix superfluous intermediate group "value" in HDF5 output.