observables/…/wavevector: fix bug with multimap keys
The comparison predicate less_tol() used to sort wavevectors by their
magnitude did not induce a strict weak ordering. The tolerance was
simply chosen too large since the stored keys are taken from the
wavenumber grid provided (and not from the magnitude of the wavevectors
found). The issue was introduced in commit 1d73ef8b.

The issue has been detected by a failing assertion in a test run:
http://my.cdash.org/testDetails.php?test=11507122&build=493412