observables/…/correlation: change signature of compute()
Let the correlation function take care of accumulating the results.
Instead of returning the result of a single call, it receives a
reference to the accumulated results. Thereby, also non-arithmetic
result types (e.g., arrays) can be supported.

Replace compute() by operator() to emphasise the functor role that the
TCF classes play.