observables/dynamics: add writer() to TCF modules
The writer function is called by blocking_scheme. It opens the specified
data group on the file and returns a writer instance. This provides a
means to the correlation function of storing additional data, e.g., the
wavenumber.

Drop attribute 'group' of correlation functions, which has become out of
use. Rename "profiler description" in the documentation to "module
description".

User-defined correlations: rename constructor parameter 'group' to
'location', and let the user provide the full path, i.e., including the
root group 'dynamics'. The passed value serves as default for writer().

Revise Lua script of dynamics.blocking_scheme. Store mean value of the
result in "value" to comply with the H5MD standard.