observables: add intermediate scattering function S(k,t)
The module assumes that S(k,t) depends only on the magnitude of the
wavevector, which holds for an isotropic system. The average over
wavevectors with comparable magnitude is returned.

This is the first example for a rank-1 time correlation function, i.e.,
a correlation function that depends on a parameter (the wavenumber).