forces: add tabulated_external module
This module provides the possibility to add an external force by
precalculating the potential on a uniform lattice and then using an
interpolation scheme. If the external potential initially consists of
many particles it should reduce the computational cost of force
calculation.

The interpolation scheme is a functor that will be given as a
template argument to the tabulated_force module -- analogue to the
smoothing functors for pair_trunc. The only example (for now) is the
previously added cubic_hermite interpolation scheme.

Because the shear stress tensor is not interpolated for now, all derived
values (as the shear viscosity) will be incorrect.