lua: support 'memory' parameter for particle and Lennard-Jones
The memory parameter can enforce the variant of the particle module.
Previously, if there was a GPU available, it was used automatically and
the only possibility was to deactivate it via '--disable-gpu' parameter
at the halmd command line. With the optional 'memory' parameter this can
be controlled now at a script level.

Furthermore, make the Lennard-Jones potential module aware of the
attribute of the particle instance.