lua/phase_space: improve support 'memory' parameter
The memory option for potential added in 27574e5 introduced a problem
with the phase_space module, as the phase_space sample can be in gpu or
in host memory, too. But if the particle instance is in host memory, the
phase_space sample can be in the host only. Therefore, restrict the
possible phase_space sample instances depending on the memory variant of
the host.

Also add an optional memory parameter to phase_space:reader() and fix a
minor problem with the docuemnted type of some parameters of the reader.