observables/phase_space: add box data to H5MD particles group
Currently, box data are written in the time-dependent format. For a
static box this leads to the repeated storage of the same edge and
offset data.

Whether the box changes in time or not depends on the user script and on
the integrator chosen (e.g., constant volume or constant pressure). A
possible robust solution would be to select upon creation of the box
whether it will change in time or not.

Adjust the continuation tests since "box/value" does not exist.

This commit refs #198804.