observables.thermodynamics: adjust output to H5MD spec
The dimension, particle number, and box volume are always stored in a
thermodynamics group. Use of the H5MD module "thermodynamics" has to be
declared in the metadata.

Rename nparticle() to particle_number() to reflect the H5MD name of the
field.

Bind data slot volume() to Lua. Do not query mdsim.box as the volume of
a particle group may be smaller than that of the whole box.

This commit refs #198804.