thermodynamics: remove evaluation of hypervirial
The hypervirial can be used to determine the compressibility. The
results, however, suffer from poor statistics and a superior method is
provided by extrapolation of the static structure factor to zero
wavenumber.

The method has been marked in commit:deec62b as "to be deprecated" for
the next release (after 0.3). Dropping it now is actually not a good
habit, but the quantity was not used by anybody so far to my knowledge.