packages.mk: fix lib path in GCC prerequisites
If libdir is omitted during configuration, it is either lib or lib64
depending on the architecture. Then packages depending on GMP, PPL, ...
do not find the respective libraries.

The issue appeared on OpenSuSE 12.3 and may be related to
http://gcc.gnu.org/bugzilla/show_bug.cgi?id=53798.

In cloog-ppl, adding linkage against libm.a resolved the following
error:

ld: .../.ppl-0.11.2/lib/libppl.a(Init.o): undefined reference to symbol 'fegetround@@GLIBC_2.2.5'
ld: note: 'fegetround@@GLIBC_2.2.5' is defined in DSO /lib64/libm.so.6 so
try adding it to the linker command line
/lib64/libm.so.6: could not read symbols: Invalid operation