add test modularity/two_particles
The purpose of the test is to check the usability of several particle
instances. The scripts setup, equilibrate, and simulate a binary mixture
where each species is kept in a separate particle instance.

Creating particle instances from subsets of particle is not yet
possible. Even a workaround in Lua via observables.phase_space is
prohibited since the data slot returned by acquire() can not be called
from Lua. Another possible workaround would use {set,get}_{position,…}
and extract the data (nested tables!) by hand. (Which I'm not going to
implement in Lua.)

Actual checks of the final result are missing though.

This commit refs #198261.