remove dependency of potentials on mdsim::particle
C++ classes: drop explicit specification of number of species types in
the constructor. These numbers are implied by the shape of the parameter
matrices.

Check matching parameter matrices and number of particle species in the
constructor of the force classes.

Lua: provide optional parameters "species", "memory", and "label"
instead of the dependency on an instance of mdsim::particle.