packages.mk: disable parallel build for HDF5, add ninja
By default, HDF5 is now configured without the --enable-parallel flag,
as it requires MPI to be installed which is not required by HALMD. It
can be enabled by "ENABLE_PARALLEL_HDF5=1", though.

Further, add ninja - an alternative to GNU make to the packages. In
order to generate ninja build files with cmake, use
  cmake -GNinja <source-dir>