mdsim.particle: revise logic for update_force()
Rename aux_valid to aux_enabled, which reflects that fact that the
auxiliary variables are enabled for computation. This overlaps with the
meaning of aux_flag which is dropped.

Introduce aux_dirty_ to indicate that the caches for the auxiliary
variables are invalid (dirty). This has been the second meaning of
aux_valid_ in the previous implementation (actually its negation).

Now aux_dirty_ tells if the variables can be used or need an update,
while aux_enabled_ is queried by the force modules whether these
variables shall be computed.

Unify update_force_() and update_aux_() by adding the flag "with_aux" to
update_force_.