random: revise singleton module, add logging output
Before, it was not possible to obtain the RNG seed and to restart a
simulation with a certain seed.

Disable default instantiation of RNGs.

Drop the Lua function seed(). There is no point in shadowing the
function of the C++ module and assigning the same seed for host and GPU
instances (they use different RNG algorithms anyway).