forces: fix issue with check_cache()
Marking the force cache as dirty must be independent of the state of
aux_enable_. In the velocity-Verlet algorithm, the force is read twice: for the
initial positions and for the new positions. The first read uses the result of
the force computation from the previous step, only the second one needs a new
computation. The aux_enable_ flag, however, is set before the first integration
step, but is meant to be active in the last (and only) force update
only.

Here are buggy examples for a "normal" step and a "sampling" step (with
aux_enabled). In step #10, the first force update is unneeded and was
triggered by aux_enabled().

 [TRACE] performing MD step #9
 [TRACE] verlet_nvt_boltzmann: update positions and velocities
 [TRACE] verlet_nvt_boltzmann: update velocities
 [TRACE] request force
 [TRACE] compute maximum squared displacement
 [TRACE] lennard_jones: compute forces

 [TRACE] performing MD step #10
 [TRACE] enable computation of auxiliary variables
 [TRACE] verlet_nvt_boltzmann: update positions and velocities
 [TRACE] request force and auxiliary variables                      <== unneeded
 [TRACE] compute maximum squared displacement
 [TRACE] lennard_jones: compute forces with auxiliary variables
 [TRACE] verlet_nvt_boltzmann: update velocities
 [TRACE] request force
 [TRACE] compute maximum squared displacement
 [TRACE] lennard_jones: compute forces
 [TRACE] accumulator (Averaged total energy): acquire sample
 [WARNING] auxiliary variables inactive in prior force computation, use aux_enable()
 [TRACE] request force and auxiliary variables
 [TRACE] compute maximum squared displacement
 [TRACE] lennard_jones: compute forces with auxiliary variables
 [TRACE] thermodynamics (all): acquire potential energy
 [TRACE] thermodynamics (all): acquire kinetic energy

To summarise, we must ensure that aux_enabled_ is not consumed
prematurely. This reveals an issue with multi-step integrators (which we
don't have yet). NPT integrators (which we will have soon) need the
pressure in both(?) half-steps and should call enable_aux() directly.
Maybe it would be better to connect enable_aux to on_finalize() and drop
sampler:on_prepare().