implement GID modulation of density profiles
For grazing incidence diffraction experiments, the scattered intensity
is given by the correlation of weighted density fluctuations such that
fluctuations deep from within the sample do not contribute. For details,
see Dietrich and Hase, Phys. Rep. 260, 1 (1995).

Here, we introduce "modulation factors" for the computation of density
modes. These factors are implemented as functors passed as template
parameters (for the functional form) and in the constructor (for the
parameters). Thereby, the evaluation of the usual density modes is as
efficient as before.