fix bug with definition of macro HALMD_WITH_GPU
The macro was originally set in halmd/config.hpp only via #cmakedefine.
Testing the macro in halmd/version.h made it necessary to query the
macro there again. The solution was a bit unfortunate as it always
undefined the macro at the end.

Drop #cmakedefine and define HALMD_WITH_GPU in the CMake files globally,
which appears more robust.

The bug has been introduced in the previous commit.