fix bug with GID modulation of density modes
The computation of density modes is based on "unwrapped" particle
positions extended outside of the periodic box. Fix it by providing the
box height (along the last dimension only) to the modulation functors
and to let them fold the positions back. This solution is not very
efficient and is considered an ugly hack.

Note that the bug fix is not needed downstream because density modes are
computed from the wrapped particle positions.